Reorganise HUD class structure + fix HUD compare method
